move DecodingStrategies used by ApodiniREST to ApodiniVaporSupport
